% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_auth.R
\name{binxr_config_futures}
\alias{binxr_config_futures}
\title{Create binxr futures config}
\usage{
binxr_config_futures(
  api_key = Sys.getenv("BINX_API_KEY"),
  secret_key = Sys.getenv("BINX_SECRET_KEY"),
  base = "https://fapi.binance.com",
  recvWindow = 10000
)
}
\arguments{
\item{api_key}{character API key.}

\item{secret_key}{character Secret key.}

\item{base}{character Base URL for Binance Futures API.}

\item{recvWindow}{integer recvWindow in milliseconds.}
}
\value{
list Configuration list for futures requests.
}
\description{
Create binxr futures config
}
