% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{cancel_fapi_trade_order}
\alias{cancel_fapi_trade_order}
\title{Cancel a futures order}
\usage{
cancel_fapi_trade_order(
  symbol,
  orderId = NULL,
  origClientOrderId = NULL,
  config = binxr_config_futures()
)
}
\arguments{
\item{symbol}{character Trading pair symbol, e.g., "ETHUSDT".}

\item{orderId}{numeric|NULL Exchange order ID.}

\item{origClientOrderId}{character|NULL Client order ID.}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON response.
}
\description{
Cancel a futures order
}
