% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{cancel_fapi_trade_orders_all}
\alias{cancel_fapi_trade_orders_all}
\title{Cancel all open futures orders for a symbol}
\usage{
cancel_fapi_trade_orders_all(symbol, config = binxr_config_futures())
}
\arguments{
\item{symbol}{character Trading pair symbol, e.g., "ETHUSDT".}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON response.
}
\description{
Cancel all open futures orders for a symbol
}
