% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_cancel_all_algo_orders}
\alias{futures_cancel_all_algo_orders}
\title{Cancel all open Binance Futures algo orders}
\usage{
futures_cancel_all_algo_orders(symbol, config = config_futures())
}
\arguments{
\item{symbol}{Trading pair symbol.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Cancel all open Binance Futures algo orders
}
