% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_cancel_all_orders}
\alias{futures_cancel_all_orders}
\alias{cancel_fapi_trade_orders_all}
\title{Cancel all open Binance Futures orders for a symbol}
\usage{
futures_cancel_all_orders(symbol, config = config_futures())

cancel_fapi_trade_orders_all(symbol, config = binxr_config_futures())
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Cancel all open Binance Futures orders for a symbol
}
