% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_cancel_batch_orders}
\alias{futures_cancel_batch_orders}
\title{Cancel multiple Binance Futures orders}
\usage{
futures_cancel_batch_orders(
  symbol,
  order_ids = NULL,
  orig_client_order_ids = NULL,
  config = config_futures()
)
}
\arguments{
\item{symbol}{Trading pair symbol.}

\item{order_ids}{Optional numeric vector of exchange order IDs.}

\item{orig_client_order_ids}{Optional character vector of client order IDs.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Cancel multiple Binance Futures orders
}
