% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_cancel_order}
\alias{futures_cancel_order}
\alias{cancel_fapi_trade_order}
\title{Cancel a Binance Futures order}
\usage{
futures_cancel_order(
  symbol,
  order_id = NULL,
  orig_client_order_id = NULL,
  config = config_futures()
)

cancel_fapi_trade_order(
  symbol,
  orderId = NULL,
  origClientOrderId = NULL,
  config = binxr_config_futures()
)
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{order_id}{Optional exchange order ID.}

\item{orig_client_order_id}{Optional client order ID.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Cancel a Binance Futures order
}
