% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-account.R
\name{futures_close_user_data_stream}
\alias{futures_close_user_data_stream}
\title{Close a Binance Futures user data stream}
\usage{
futures_close_user_data_stream(listen_key, config = config_futures())
}
\arguments{
\item{listen_key}{Listen key returned by \code{futures_start_user_data_stream()}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}} with an API key.}
}
\value{
A parsed list.
}
\description{
Close a Binance Futures user data stream
}
