% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_countdown_cancel_all}
\alias{futures_countdown_cancel_all}
\title{Set Binance Futures countdown cancel-all}
\usage{
futures_countdown_cancel_all(symbol, countdown_time, config = config_futures())
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{countdown_time}{Countdown time in milliseconds.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Set Binance Futures countdown cancel-all
}
