% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-account.R
\name{futures_get_account}
\alias{futures_get_account}
\alias{get_fapi_account}
\title{Get Binance Futures account info}
\usage{
futures_get_account(config = config_futures())

get_fapi_account(config = binxr_config_futures())
}
\arguments{
\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures account info
}
