% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-account.R
\name{futures_get_account_config}
\alias{futures_get_account_config}
\title{Get Binance Futures account configuration}
\usage{
futures_get_account_config(config = config_futures())
}
\arguments{
\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures account configuration
}
