% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_get_algo_order}
\alias{futures_get_algo_order}
\title{Get a Binance Futures algo order}
\usage{
futures_get_algo_order(
  symbol,
  algo_id = NULL,
  client_algo_id = NULL,
  config = config_futures()
)
}
\arguments{
\item{symbol}{Trading pair symbol.}

\item{algo_id}{Optional exchange algo-order ID.}

\item{client_algo_id}{Optional client algo-order ID.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get a Binance Futures algo order
}
