% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-account.R
\name{futures_get_api_trading_status}
\alias{futures_get_api_trading_status}
\title{Get Binance Futures API trading status}
\usage{
futures_get_api_trading_status(symbol = NULL, config = config_futures())
}
\arguments{
\item{symbol}{Optional trading pair symbol.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures API trading status
}
