% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-market.R
\name{futures_get_book_ticker}
\alias{futures_get_book_ticker}
\title{Get Binance Futures order book ticker}
\usage{
futures_get_book_ticker(symbol = NULL, config = config_futures())
}
\arguments{
\item{symbol}{Optional trading pair symbol.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures order book ticker
}
