% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-market.R
\name{futures_get_index_constituents}
\alias{futures_get_index_constituents}
\title{Get Binance Futures index constituents}
\usage{
futures_get_index_constituents(symbol, config = config_futures())
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures index constituents
}
