% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-account.R
\name{futures_get_multi_assets_mode}
\alias{futures_get_multi_assets_mode}
\title{Get Binance Futures multi-assets mode}
\usage{
futures_get_multi_assets_mode(config = config_futures())
}
\arguments{
\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures multi-assets mode
}
