% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_get_order}
\alias{futures_get_order}
\alias{get_fapi_trade_order}
\title{Get a Binance Futures order}
\usage{
futures_get_order(
  symbol,
  order_id = NULL,
  orig_client_order_id = NULL,
  config = config_futures()
)

get_fapi_trade_order(
  symbol,
  orderId = NULL,
  origClientOrderId = NULL,
  config = binxr_config_futures()
)
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{order_id}{Optional exchange order ID.}

\item{orig_client_order_id}{Optional client order ID.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get a Binance Futures order
}
