% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-market.R
\name{futures_get_symbol_adl_risk}
\alias{futures_get_symbol_adl_risk}
\title{Get Binance Futures symbol-level ADL risk}
\usage{
futures_get_symbol_adl_risk(symbol, config = config_futures())
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures symbol-level ADL risk
}
