% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-market.R
\name{futures_get_ticker_price}
\alias{futures_get_ticker_price}
\title{Get Binance Futures ticker price}
\usage{
futures_get_ticker_price(
  symbol = NULL,
  version = c("v2", "deprecated"),
  config = config_futures()
)
}
\arguments{
\item{symbol}{Optional trading pair symbol.}

\item{version}{One of \code{"v2"} or \code{"deprecated"}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures ticker price
}
