% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-market.R
\name{futures_get_trading_schedule}
\alias{futures_get_trading_schedule}
\title{Get Binance Futures trading schedules}
\usage{
futures_get_trading_schedule(config = config_futures())
}
\arguments{
\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Futures trading schedules
}
