% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_modify_position_margin}
\alias{futures_modify_position_margin}
\title{Change Binance Futures position margin}
\usage{
futures_modify_position_margin(
  symbol,
  amount,
  type = c("ADD", "REDUCE"),
  position_side = c("BOTH", "LONG", "SHORT"),
  config = config_futures()
)
}
\arguments{
\item{symbol}{Trading pair symbol.}

\item{amount}{Margin amount to add or reduce.}

\item{type}{Either \code{"ADD"} or \code{"REDUCE"}.}

\item{position_side}{One of \code{"BOTH"}, \code{"LONG"}, or \code{"SHORT"}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Change Binance Futures position margin
}
