% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-market.R
\name{futures_ping}
\alias{futures_ping}
\title{Test Binance Futures connectivity}
\usage{
futures_ping(config = config_futures())
}
\arguments{
\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Test Binance Futures connectivity
}
