% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_set_leverage}
\alias{futures_set_leverage}
\alias{set_fapi_account_leverage}
\title{Set Binance Futures leverage}
\usage{
futures_set_leverage(symbol, leverage, config = config_futures())

set_fapi_account_leverage(symbol, leverage, config = binxr_config_futures())
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{leverage}{Desired leverage.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Set Binance Futures leverage
}
