% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_set_margin_type}
\alias{futures_set_margin_type}
\alias{set_fapi_account_margin_type}
\title{Set Binance Futures margin type}
\usage{
futures_set_margin_type(
  symbol,
  margin_type = c("CROSSED", "ISOLATED"),
  config = config_futures()
)

set_fapi_account_margin_type(
  symbol,
  marginType = c("CROSSED", "ISOLATED"),
  config = binxr_config_futures()
)
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"ETHUSDT"}.}

\item{margin_type}{One of \code{"CROSSED"} or \code{"ISOLATED"}.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Set Binance Futures margin type
}
