% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_set_multi_assets_mode}
\alias{futures_set_multi_assets_mode}
\title{Set Binance Futures multi-assets mode}
\usage{
futures_set_multi_assets_mode(
  multi_assets_margin = TRUE,
  config = config_futures()
)
}
\arguments{
\item{multi_assets_margin}{\code{TRUE} to enable multi-assets mode, \code{FALSE} to disable it.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Set Binance Futures multi-assets mode
}
