% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-trade.R
\name{futures_set_position_mode}
\alias{futures_set_position_mode}
\alias{set_fapi_account_position_side}
\title{Set Binance Futures position mode}
\usage{
futures_set_position_mode(dual_side_position = TRUE, config = config_futures())

set_fapi_account_position_side(
  dualSidePosition = TRUE,
  config = binxr_config_futures()
)
}
\arguments{
\item{dual_side_position}{\code{TRUE} for hedge mode, \code{FALSE} for one-way mode.}

\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}}.}
}
\value{
A parsed list.
}
\description{
Set Binance Futures position mode
}
