% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-futures-account.R
\name{futures_start_user_data_stream}
\alias{futures_start_user_data_stream}
\title{Start a Binance Futures user data stream}
\usage{
futures_start_user_data_stream(config = config_futures())
}
\arguments{
\item{config}{A futures configuration created by \code{\link[=config_futures]{config_futures()}} with an API key.}
}
\value{
A parsed list containing a listen key.
}
\description{
Start a Binance Futures user data stream
}
