% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_fapi_account}
\alias{get_fapi_account}
\title{Get Binance Futures account info}
\usage{
get_fapi_account(config = binxr_config_futures())
}
\arguments{
\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON from \verb{/fapi/v2/account}.
}
\description{
Get Binance Futures account info
}
