% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_fapi_account_position_risk}
\alias{get_fapi_account_position_risk}
\title{Get position risk (Futures)}
\usage{
get_fapi_account_position_risk(config = binxr_config_futures())
}
\arguments{
\item{config}{list binxr futures configuration.}
}
\value{
data.table Position risk table (long positions only) with \code{updateTime} as POSIXct; returned invisibly.
}
\description{
Get position risk (Futures)
}
