% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_fapi_account_positions}
\alias{get_fapi_account_positions}
\title{Get standardized account positions (Futures)}
\usage{
get_fapi_account_positions(acc = NULL, config = binxr_config_futures())
}
\arguments{
\item{acc}{list|NULL Account object from \code{get_fapi_account()}; if NULL it will be fetched.}

\item{config}{list binxr futures configuration.}
}
\value{
data.table Positions with numeric columns coerced and \code{updateTime} as POSIXct; returned invisibly.
}
\description{
Get standardized account positions (Futures)
}
