% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_fapi_account_summary}
\alias{get_fapi_account_summary}
\title{Get account summary (Futures)}
\usage{
get_fapi_account_summary(acc = NULL, config = binxr_config_futures())
}
\arguments{
\item{acc}{list|NULL Account object from \code{get_fapi_account()}; if NULL it will be fetched.}

\item{config}{list binxr futures configuration.}
}
\value{
data.table One-row summary with numeric balances and margins; returned invisibly.
}
\description{
Get account summary (Futures)
}
