% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{get_fapi_exchange_info}
\alias{get_fapi_exchange_info}
\title{Get Binance Futures exchange info}
\usage{
get_fapi_exchange_info(config = binxr_config_futures())
}
\arguments{
\item{config}{list binxr futures configuration.}
}
\value{
list Exchange info (symbols, assets); returned invisibly.
}
\description{
Get Binance Futures exchange info
}
