% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{get_fapi_mark_price}
\alias{get_fapi_mark_price}
\title{Get Binance Futures mark price (premium index)}
\usage{
get_fapi_mark_price(symbol = NULL, config = binxr_config_futures())
}
\arguments{
\item{symbol}{character|NULL Trading pair symbol like "ETHUSDT"; NULL for all.}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON from \verb{/fapi/v1/premiumIndex}.
}
\description{
Get Binance Futures mark price (premium index)
}
