% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{get_fapi_system_time}
\alias{get_fapi_system_time}
\title{Get Binance Futures server time}
\usage{
get_fapi_system_time(config = binxr_config_futures())
}
\arguments{
\item{config}{list binxr futures configuration.}
}
\value{
POSIXct Current server time.
}
\description{
Get Binance Futures server time
}
