% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{get_fapi_trade_open_orders}
\alias{get_fapi_trade_open_orders}
\title{Get open orders (Futures)}
\usage{
get_fapi_trade_open_orders(symbol = NULL, config = binxr_config_futures())
}
\arguments{
\item{symbol}{character|NULL Trading pair symbol; NULL for all.}

\item{config}{list binxr futures configuration.}
}
\value{
data.table Open orders with \code{time_human} and \code{updateTime_human} columns.
}
\description{
Get open orders (Futures)
}
