% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market.R
\name{get_spot_mark_price}
\alias{get_spot_mark_price}
\title{Get Binance Spot mark price}
\usage{
get_spot_mark_price(symbol = NULL, config = binxr_config_spot())
}
\arguments{
\item{symbol}{character|NULL Trading pair symbol like "ETHUSDT"; NULL for all.}

\item{config}{list binxr spot configuration.}
}
\value{
list Parsed JSON from \verb{/api/v3/ticker/price}.
}
\description{
Get Binance Spot mark price
}
