% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-trade.R
\name{options_cancel_batch_orders}
\alias{options_cancel_batch_orders}
\title{Cancel multiple Binance Options orders}
\usage{
options_cancel_batch_orders(order_ids, config = config_options())
}
\arguments{
\item{order_ids}{A non-empty numeric vector of exchange order IDs.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Cancel multiple Binance Options orders
}
