% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-market-maker.R
\name{options_cancel_block_trade_order}
\alias{options_cancel_block_trade_order}
\title{Cancel a Binance Options block trade order}
\usage{
options_cancel_block_trade_order(
  block_order_matching_key,
  config = config_options()
)
}
\arguments{
\item{block_order_matching_key}{Block trade matching key.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Cancel a Binance Options block trade order
}
