% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-trade.R
\name{options_cancel_order}
\alias{options_cancel_order}
\title{Cancel a Binance Options order}
\usage{
options_cancel_order(
  symbol,
  order_id = NULL,
  client_order_id = NULL,
  config = config_options()
)
}
\arguments{
\item{symbol}{Option symbol, for example \code{"BTC-200730-9000-C"}.}

\item{order_id}{Optional exchange order ID.}

\item{client_order_id}{Optional client order ID.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Cancel a Binance Options order
}
