% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-account.R
\name{options_close_user_data_stream}
\alias{options_close_user_data_stream}
\title{Close a Binance Options user data stream}
\usage{
options_close_user_data_stream(listen_key, config = config_options())
}
\arguments{
\item{listen_key}{Listen key returned by \code{options_start_user_data_stream()}.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}} with an API key.}
}
\value{
A parsed list.
}
\description{
Close a Binance Options user data stream
}
