% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-market-maker.R
\name{options_get_block_trade_details}
\alias{options_get_block_trade_details}
\title{Get Binance Options block trade details}
\usage{
options_get_block_trade_details(
  block_order_matching_key,
  config = config_options()
)
}
\arguments{
\item{block_order_matching_key}{Block trade matching key.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Options block trade details
}
