% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-market.R
\name{options_get_index_price}
\alias{options_get_index_price}
\title{Get Binance Options underlying index price}
\usage{
options_get_index_price(underlying, config = config_options())
}
\arguments{
\item{underlying}{Underlying spot pair, for example \code{"BTCUSDT"}.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Options underlying index price
}
