% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-account.R
\name{options_get_margin_account}
\alias{options_get_margin_account}
\title{Get Binance Options margin account information}
\usage{
options_get_margin_account(config = config_options())
}
\arguments{
\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Options margin account information
}
