% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-market-maker.R
\name{options_get_market_maker_protection}
\alias{options_get_market_maker_protection}
\title{Get Binance Options market maker protection configuration}
\usage{
options_get_market_maker_protection(underlying, config = config_options())
}
\arguments{
\item{underlying}{Underlying symbol, for example \code{"BTCUSDT"}.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
This endpoint is available only to eligible Options market-maker accounts.
}
