% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-market.R
\name{options_get_order_book}
\alias{options_get_order_book}
\title{Get Binance Options order book}
\usage{
options_get_order_book(symbol, limit = NULL, config = config_options())
}
\arguments{
\item{symbol}{Option symbol, for example \code{"BTC-200730-9000-C"}.}

\item{limit}{Optional depth limit. One of \code{10}, \code{20}, \code{50}, \code{100}, \code{500}, or
\code{1000}.}

\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Options order book
}
