% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-trade.R
\name{options_sign_stock_contract}
\alias{options_sign_stock_contract}
\title{Sign the Binance Options stock contract}
\usage{
options_sign_stock_contract(config = config_options())
}
\arguments{
\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}}.}
}
\value{
A parsed list.
}
\description{
Sign the Binance Options stock contract
}
