% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-options-account.R
\name{options_start_user_data_stream}
\alias{options_start_user_data_stream}
\title{Start a Binance Options user data stream}
\usage{
options_start_user_data_stream(config = config_options())
}
\arguments{
\item{config}{An options configuration created by \code{\link[=config_options]{config_options()}} with an API key.}
}
\value{
A parsed list containing a listen key.
}
\description{
Start a Binance Options user data stream
}
