% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{place_fapi_trade_order}
\alias{place_fapi_trade_order}
\title{Place futures trade order}
\usage{
place_fapi_trade_order(
  symbol,
  side = c("BUY", "SELL"),
  type = c("LIMIT", "MARKET", "STOP", "STOP_MARKET", "TAKE_PROFIT", "TAKE_PROFIT_MARKET",
    "TRAILING_STOP_MARKET"),
  quantity,
  price = NULL,
  timeInForce = c("GTC", "IOC", "FOK", "GTX"),
  positionSide = c("BOTH", "LONG", "SHORT"),
  reduceOnly = FALSE,
  stopPrice = NULL,
  workingType = c("CONTRACT_PRICE", "MARK_PRICE"),
  newClientOrderId = NULL,
  config = binxr_config_futures()
)
}
\arguments{
\item{symbol}{character Trading pair symbol, e.g., "ETHUSDT".}

\item{side}{character "BUY" or "SELL".}

\item{type}{character Order type, e.g., "LIMIT", "MARKET", "STOP", etc.}

\item{quantity}{numeric Order quantity.}

\item{price}{numeric|NULL Limit price for limit orders.}

\item{timeInForce}{character One of "GTC","IOC","FOK","GTX" (for limit).}

\item{positionSide}{character "BOTH","LONG","SHORT".}

\item{reduceOnly}{logical Whether the order is reduce-only.}

\item{stopPrice}{numeric|NULL Stop/trigger price for conditional orders.}

\item{workingType}{character "CONTRACT_PRICE" or "MARK_PRICE".}

\item{newClientOrderId}{character|NULL Optional client order ID.}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON response.
}
\description{
Place futures trade order
}
