% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{set_fapi_account_leverage}
\alias{set_fapi_account_leverage}
\title{Set leverage (Futures)}
\usage{
set_fapi_account_leverage(symbol, leverage, config = binxr_config_futures())
}
\arguments{
\item{symbol}{character Trading pair symbol, e.g., "ETHUSDT".}

\item{leverage}{integer Desired leverage.}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON response.
}
\description{
Set leverage (Futures)
}
