% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{set_fapi_account_margin_type}
\alias{set_fapi_account_margin_type}
\title{Set margin type (Futures)}
\usage{
set_fapi_account_margin_type(
  symbol,
  marginType = c("CROSSED", "ISOLATED"),
  config = binxr_config_futures()
)
}
\arguments{
\item{symbol}{character Trading pair symbol, e.g., "ETHUSDT".}

\item{marginType}{character One of "CROSSED" or "ISOLATED".}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON response.
}
\description{
Set margin type (Futures)
}
