% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\name{set_fapi_account_position_side}
\alias{set_fapi_account_position_side}
\title{Set position mode (dual-side) (Futures)}
\usage{
set_fapi_account_position_side(
  dualSidePosition = TRUE,
  config = binxr_config_futures()
)
}
\arguments{
\item{dualSidePosition}{logical TRUE for dual-side, FALSE for one-way.}

\item{config}{list binxr futures configuration.}
}
\value{
list Parsed JSON response.
}
\description{
Set position mode (dual-side) (Futures)
}
