% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-spot-general.R
\name{spot_get_execution_rules}
\alias{spot_get_execution_rules}
\title{Get Binance Spot execution rules}
\usage{
spot_get_execution_rules(
  symbol = NULL,
  symbols = NULL,
  symbol_status = NULL,
  config = config_spot()
)
}
\arguments{
\item{symbol}{Optional trading pair symbol.}

\item{symbols}{Optional character vector of trading pair symbols.}

\item{symbol_status}{Optional trading status filter.}

\item{config}{A spot configuration created by \code{\link[=config_spot]{config_spot()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Spot execution rules
}
