% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-spot-account.R
\name{spot_get_order}
\alias{spot_get_order}
\title{Get a Binance Spot order}
\usage{
spot_get_order(
  symbol,
  order_id = NULL,
  orig_client_order_id = NULL,
  config = config_spot()
)
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"BTCUSDT"}.}

\item{order_id}{Optional exchange order ID.}

\item{orig_client_order_id}{Optional client order ID.}

\item{config}{A spot configuration created by \code{\link[=config_spot]{config_spot()}}.}
}
\value{
A parsed list.
}
\description{
Get a Binance Spot order
}
