% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-spot-market.R
\name{spot_get_reference_price_calculation}
\alias{spot_get_reference_price_calculation}
\title{Get Binance Spot reference price calculation metadata}
\usage{
spot_get_reference_price_calculation(
  symbol,
  symbol_status = NULL,
  config = config_spot()
)
}
\arguments{
\item{symbol}{Trading pair symbol, for example \code{"BTCUSDT"}.}

\item{symbol_status}{Optional trading status filter.}

\item{config}{A spot configuration created by \code{\link[=config_spot]{config_spot()}}.}
}
\value{
A parsed list.
}
\description{
Get Binance Spot reference price calculation metadata
}
