% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints-spot-general.R
\name{spot_get_server_time}
\alias{spot_get_server_time}
\title{Get Binance Spot server time}
\usage{
spot_get_server_time(config = config_spot())
}
\arguments{
\item{config}{A spot configuration created by \code{\link[=config_spot]{config_spot()}}.}
}
\value{
A \code{POSIXct} timestamp.
}
\description{
Get Binance Spot server time
}
