% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{coerce_account_frames}
\alias{coerce_account_frames}
\title{(Optional) Coerce account to data.frames}
\usage{
coerce_account_frames(account)
}
\arguments{
\item{account}{get_account() result}
}
\value{
list(assets=..., positions=...)
}
\description{
(Optional) Coerce account to data.frames
}
